// js/inc/back/attempt02_main/lessons_my/ls03/src/drivers/dto/driver.input-dto.ts
import { VehicleFeature } from "../types/driver";

export type DriverInputDto = {
  name: string;
  phoneNumber: string;
  email: string;
  vehicleMake: string;
  vehicleModel: string;
  vehicleYear: number;
  vehicleLicensePlate: string;
  vehicleDescription: string | null;
  vehicleFeatures: VehicleFeature[];
};